% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hommen212}
\alias{hommen212}
\title{Observed frond numbers reported by Hommen \emph{et al.} (2015)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25 rows and 3 columns.
}
\usage{
hommen212
}
\description{
The data consists of observed frond numbers from experimental studies for
various time points and exposure concentrations. The \emph{Lemna} population
was exposed to constant concentrations of \emph{metsulfuron-methyl} for two days,
followed by twelve days of recovery.
}
\details{
The dataset was presented in Hommen \emph{et al.} (2015)],
cf. Figure 3, and was included in this package by courtesy of the authors.
}
\section{Literature}{
\itemize{
\item Hommen U, Schmitt W, Heine S, Brock Theo CM, Duquesne S, Manson P, Meregalli G,
Ochoa-Acuña H, van Vliet P, Arts G. 2015. How TK-TD and Population Models for
Aquatic Macrophytes Could Support the Risk Assessment for Plant Protection
Products. Integr Environ Assess Manag 12(1), pp. 82-95.
\doi{10.1002/ieam.1715}
}
}

\keyword{datasets}
