% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemna.R
\name{lemna_desolve}
\alias{lemna_desolve}
\title{Full access to ODE solver}
\usage{
lemna_desolve(nout = 2, ...)
}
\arguments{
\item{nout}{optional \code{numerical}, number of additional output variables,
defaults to two}

\item{...}{additional parameters passed on to \code{\link[deSolve:ode]{deSolve::ode()}}}
}
\value{
result from \code{\link[deSolve:ode]{deSolve::ode()}}
}
\description{
This function can be used by external packages to access the ODE implemented
in C without the surrounding sanity checks and data loading procedures.
All parameters will be passed on to the solver. This structure may help
avoiding CMD CHECK messages.
}
