% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemna_scenario.R
\name{new_lemna_scenario}
\alias{new_lemna_scenario}
\title{Lemna scenario constructor}
\usage{
new_lemna_scenario(
  init = c(BM = 0, M_int = 0),
  times = c(),
  param = list(),
  envir = list()
)
}
\arguments{
\item{init}{initial state of the model variables}

\item{times}{numeric vector, output times for which model results are returned}

\item{param}{named list, Lemna model parameters}

\item{envir}{named list, contains time series data for each of the five
environmental variables}
}
\value{
a \code{lemna_scenario} object
}
\description{
Lemna scenario constructor
}
