% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schmitt77}
\alias{schmitt77}
\title{Observed frond numbers reported by Schmitt \emph{et al.} (2013)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49 rows and 3 columns.
}
\usage{
schmitt77
}
\description{
The data consists of observed frond numbers from experimental studies for
various time points and exposure concentrations. The \emph{Lemna} population
was exposed to constant concentrations of \emph{metsulfuron-methyl} for seven days,
followed by seven days of recovery.
}
\details{
The dataset was presented in Schmitt \emph{et al.} (2013),
cf. Figure 2, and was included in this package by courtesy of the authors.
}
\section{Literature}{
\itemize{
\item Schmitt W, Bruns E, Dollinger M, Sowig P. 2013. Mechanistic TK/TD-model
simulating the effect of growth inhibitors on \emph{Lemna} populations. Ecol Model
255, pp. 1-10. \doi{10.1016/j.ecolmodel.2013.01.017}
}
}

\keyword{datasets}
