% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemna.R
\name{lemna_desolve}
\alias{lemna_desolve}
\title{Access to the ODE solver}
\usage{
lemna_desolve(...)
}
\arguments{
\item{...}{parameters passed on to \code{\link[deSolve:ode]{deSolve::ode()}}}
}
\value{
result from \code{\link[deSolve:ode]{deSolve::ode()}}
}
\description{
This function can be used by external packages to access the ODE implemented
in C without the surrounding sanity checks and data loading procedures.
All parameters will be passed on to the solver.
}
